# number_format: float within range of int can be cast to int …