…before rounding of numbers above 2^52 to not loose precision.

On casting float to int fractional digits gets truncated but as far as I'm aware of `double` numbers above 2^52 can't store fractions anymore.

This is the diff of the added test numbers without this change:
```
--
     --- testing: float(9.223372036854775E+18)
     ... with precision 5: string(31) "9,223,372,036,854,774,784.00000"
     ... with precision 0: string(25) "9,223,372,036,854,774,784"
169- ... with precision -1: string(25) "9,223,372,036,854,774,780"
170- ... with precision -5: string(25) "9,223,372,036,854,800,000"
169+ ... with precision -1: string(25) "9,223,372,036,854,774,784"
170+ ... with precision -5: string(25) "9,223,372,036,854,800,384"
     ... with precision -10: string(25) "9,223,372,040,000,000,000"
     ... with precision -11: string(25) "9,223,372,000,000,000,000"
     ... with precision -17: string(25) "9,200,000,000,000,000,000"
--
     --- testing: float(-9.223372036854775E+18)
     ... with precision 5: string(32) "-9,223,372,036,854,774,784.00000"
     ... with precision 0: string(26) "-9,223,372,036,854,774,784"
180- ... with precision -1: string(26) "-9,223,372,036,854,774,780"
181- ... with precision -5: string(26) "-9,223,372,036,854,800,000"
180+ ... with precision -1: string(26) "-9,223,372,036,854,774,784"
181+ ... with precision -5: string(26) "-9,223,372,036,854,800,384"
     ... with precision -10: string(26) "-9,223,372,040,000,000,000"
     ... with precision -11: string(26) "-9,223,372,000,000,000,000"
     ... with precision -17: string(26) "-9,200,000,000,000,000,000"
--
```