@marc-mabe 2023-10-02
Updated to 2^52 now.

Also a quick benchmark:
```php
<?php

$number = 4503599627370496.0;
$iterations = 100000;
$start = hrtime(true);
for ($i = 0; $i < $iterations; $i += 1) {
    $tmp = number_format($number);
}
$end = hrtime(true);
$spend = $end - $start;
printf('%s iterations: %Fs (%Fns per run)', $iterations, $spend / 1000000000, $spend / $iterations);
```

Before: `100000 iterations: 0.093053s (930.532260ns per run)`
After  :: `100000 iterations: 0.034346s (343.463090ns per run)`